Change facebookId in model to BIGINT

+++ models/artist.js
@@ -3,7 +3,7 @@
 module.exports = function(sequelize, DataTypes) {
   var Artist = sequelize.define('Artist', {
     id:         { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
-    facebookId: { type: DataTypes.INTEGER, unique: true, allowNull: false }
+    facebookId: { type: DataTypes.BIGINT, unique: true, allowNull: false }
   }, {
     timestamps: false,
     tableName: 'Artists'
